Redirect to posts on GitHub
In a minimal effort to prevent link rot this redirects any /blog/post
URLs to the Markdown file on GitHub.

It achieves this with some JavaScript in 404.html that detects that
we're on a `/blog/post` and then does some Regex magic to find the
corresponding URL in the GitHub repo.